Move helpers to common module and re-use them
