+++ flocker/testtools/test/test_base.py
@@ -6,6 +6,7 @@ import errno
 import os
 import shutil
 import string
+import unittest
 
 from hypothesis import assume, given
 from hypothesis.strategies import integers, lists, text
@@ -26,14 +27,16 @@ from testtools.matchers import (
     PathExists,
     StartsWith,
 )
-from testtools.testresult.doubles import Python27TestResult
 from twisted.python.filepath import FilePath
-from twisted.trial import unittest
 
 from .._base import (
     AsyncTestCase,
     make_temporary_directory,
     _path_for_test_id,
+)
+from .._testhelpers import (
+    only_skips,
+    run_test,
 )
 
 
@@ -54,16 +57,8 @@ class AsyncTestCaseTests(TestCase):
                 raise unittest.SkipTest(reason)
 
         test = SkippingTest('test_skip')
-        # We need a test result double that we can useful results from, and
-        # the Python 2.7 TestResult is the lowest common denominator.
-        logs = []
-        result = Python27TestResult(logs)
-        test.run(result)
-        self.assertEqual([
-            ('startTest', test),
-            ('addSkip', test, reason),
-            ('stopTest', test),
-        ], logs)
+        result = run_test(test)
+        self.assertThat(result, only_skips(1, [reason]))
 
     def test_mktemp_doesnt_exist(self):
         """
@@ -95,7 +90,7 @@ class AsyncTestCaseTests(TestCase):
                 created_files.append(path)
                 open(path, 'w').write('hello')
 
-        SomeTest('test_create_file').run(Python27TestResult())
+        run_test(SomeTest('test_create_file'))
         [path] = created_files
         self.addCleanup(os.unlink, path)
         self.assertThat(path, FileContains('hello'))
